ImGui Implementation
Again, Mouse position/Window position issues

Matoya technically gives you the window Client Area (correct), but the position of the mouse does not seem to correlate well. Without scaling, the mouse position is off by a bit in fullscreen but with small and smaller sizes, it just goes totally off

Issue I have to solve because it is something stupid probably, still my head is a bit fried.